tesis_back/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
pydantic
pydantic-settings
python-jose
passlib[bcrypt]
paho-mqtt
python-multipart
email-validator
